functions moved to tests
