+++ tests/test_oml/test_functional/test_metrics/test_retrieval_metrics.py
@@ -7,19 +7,21 @@ import torch
 from oml.functional.losses import surrogate_precision
 from oml.functional.metrics import (
     TMetricsDict,
-    apply_mask_to_ignore,
     calc_cmc,
     calc_fnmr_at_fmr,
-    calc_gt_mask,
     calc_map,
-    calc_mask_to_ignore,
     calc_precision,
-)
-from oml.functional.metrics import (
-    calc_retrieval_metrics_on_full as calc_retrieval_metrics,
 )
 from oml.utils.misc import remove_unused_kwargs
 from oml.utils.misc_torch import take_2d
+from tests.test_integrations.utils import (
+    apply_mask_to_ignore,
+    calc_gt_mask,
+    calc_mask_to_ignore,
+)
+from tests.test_integrations.utils import (
+    calc_retrieval_metrics_on_full as calc_retrieval_metrics,
+)
 
 from .synthetic import generate_distance_matrix, generate_retrieval_case
